aarch64: add mmc driver test
