+++ kernel/build.rs
@@ -5,14 +5,16 @@ use std::io::{Result, Write};
 
 fn main() {
     println!("cargo:rerun-if-env-changed=LOG");
-    println!("cargo:rerun-if-env-changed=SMP");
     println!("cargo:rerun-if-env-changed=BOARD");
-    println!("cargo:rerun-if-env-changed=USER_IMG");
 
     let arch: String = std::env::var("ARCH").unwrap();
     let _board: String = std::env::var("BOARD").unwrap();
-    if let Ok(user_img) = std::env::var("USER_IMG") {
-        println!("cargo:rerun-if-changed={}", user_img);
+
+    if cfg!(feature = "link_user") {
+        println!("cargo:rerun-if-env-changed=USER_IMG");
+        if let Ok(user_img) = std::env::var("USER_IMG") {
+            println!("cargo:rerun-if-changed={}", user_img);
+        }
     }
 
     match arch.as_str() {
@@ -22,7 +24,9 @@ fn main() {
         "riscv32" => {}
         "riscv64" => {}
         "mipsel" => {}
-        "aarch64" => {}
+        "aarch64" => {
+            println!("cargo:rerun-if-env-changed=SMP");
+        }
         _ => panic!("Unknown arch {}", arch),
     }
 }
