+++ kernel/src/drivers/gpu/virtio_gpu.rs
@@ -17,7 +17,6 @@ use crate::HEAP_ALLOCATOR;
 
 use super::super::bus::virtio_mmio::*;
 use super::super::{DeviceType, Driver, DRIVERS, IRQ_MANAGER};
-use super::test::mandelbrot;
 use crate::memory::phys_to_virt;
 
 const VIRTIO_GPU_EVENT_DISPLAY: u32 = 1 << 0;
@@ -267,11 +266,12 @@ fn setup_framebuffer(driver: &mut VirtIOGpu) {
         HEAP_ALLOCATOR.alloc_zeroed(Layout::from_size_align(size as usize, PAGE_SIZE).unwrap())
     } as usize;
     // test framebuffer
-//    mandelbrot(
-//        driver.rect.width,
-//        driver.rect.height,
-//        frame_buffer as *mut u32,
-//    );
+    // use super::test::mandelbrot;
+    // mandelbrot(
+    //     driver.rect.width,
+    //     driver.rect.height,
+    //     frame_buffer as *mut u32,
+    // );
     driver.frame_buffer = frame_buffer;
     let request_resource_attach_backing = unsafe {
         &mut *(driver.queue_buffer[VIRTIO_BUFFER_TRANSMIT] as *mut VirtIOGpuResourceAttachBacking)
