+++ kernel/src/drivers/mmc/mod.rs
@@ -1,2 +1,4 @@
+mod test;
+
 #[cfg(feature = "board_raspi3")]
 pub mod bcm2835_sdhci;
